Change some code and design

+++ M_CGPA/BLL/GradePoingBll.cs
@@ -54,7 +54,7 @@ namespace M_CGPA.BLL
             return _gradePointDal.GetAll();
         }
 
-        public DataTable GetByMark(double mark, int totalMark)
+        public DataTable GetByMark(double mark, double totalMark)
         {
             return _gradePointDal.GetByMark(mark, totalMark);
         }
